+++ tests/testthat/test-adjust_mortality.R
@@ -1,10 +1,10 @@
 library(dplyr)
 test_that("fallen dead trees get NAs correctly", {
-  db <- read_fia(
+  db <- fia_load(
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID == "10_1_1_104_1_28") |>
     dplyr::select(
       plot_ID,
@@ -37,11 +37,11 @@ test_that("fallen dead trees get NAs correctly", {
 })
 
 test_that("trees moving to non-sampled conditions have NAs", {
-  db <- read_fia(
+  db <- fia_load(
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID == "10_1_1_22_4_3") |>
     dplyr::select(
       plot_ID,
@@ -74,11 +74,11 @@ test_that("trees moving to non-sampled conditions have NAs", {
 })
 
 test_that("method doesn't matter for DE", {
-  db <- read_fia(
+  db <- fia_load(
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID == "10_1_1_104_1_28") |>
     dplyr::select(
       plot_ID,
@@ -112,11 +112,11 @@ test_that("method doesn't matter for DE", {
 })
 
 test_that("No values below thresholds for measurement", {
-  db <- read_fia(
+  db <- fia_load(
     "DE",
     dir = system.file("exdata", package = "forestTIME.builder")
   )
-  data <- prep_data(db) |>
+  data <- fia_tidy(db) |>
     dplyr::filter(tree_ID %in% c("10_1_1_104_3_4", "10_1_1_148_4_2")) |>
     dplyr::select(
       plot_ID,